requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-adk==1.0.0
google-genai==1.31.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
